Extract out new module `TokenFingerprint`.
